% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/azimuth_dashboard.R
\name{AzimuthDashboardApp}
\alias{AzimuthDashboardApp}
\title{Launch the dashboard app}
\usage{
AzimuthDashboardApp(config = NULL)
}
\arguments{
\item{config}{json file with googlesheet config info}
}
\value{
None, launches the Shiny app
}
\description{
Launch the dashboard app
}
