% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{AzimuthDashboardServer}
\alias{AzimuthDashboardServer}
\title{Server function for dashboard}
\usage{
AzimuthDashboardServer(input, output, session)
}
\arguments{
\item{input, output, session}{Required Shiny app server parameters}
}
\value{
The shiny server logic
}
\description{
Server function for dashboard
}
\keyword{internal}
